BooPickle and Circe Decode Instances: redo.
In both the `BooPickle` and the `Circe` decoder instances, a call
to `collectBinary` is followed by an analysis to extract from the
chunk the result.

This analysis phase was written as a function of Chunk to a
DecodeResult, flatMapped to the `collectBinary`.
This PR rewrites them as pure functions of `Either` to `Either`,
which are applied to the result of `collectBinary` using
the `EitherT.subflatMap` function.